requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
boto3==1.34.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
gspread==5.12.0
pandas==2.1.4
openpyxl==3.1.2
Pillow==10.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
pytesseract==0.3.10